web_application/backend/microservices/controller_task_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.6.3
pydantic-settings[yaml]==2.2.1
PyYaml==6.0.1
starlette==0.27.0
asyncpg==0.29.0
databases[postgresql]==0.2.6
alembic==1.13.1
pytest==7.4.3
fastapi-users[sqlalchemy]==12.1.2
fastapi_users_db_sqlalchemy==6.0.1
requests==2.31.0
pytest-asyncio==0.23.2
aiohttp==3.9.1
httpx==0.25.2
